tap: generalize call tapping feature
The debug feature to tap certain calls is currently very NET/BTS
oriented and also limited to two connections.

Generalize the feature in a way that it works con RTP connections
directly. The user can then tap arbitrary RTP connections on one
endpoint, no matter if it is on the network or on the BTS side.