stat: generalize statistics
When a connection is deleted (DLCX) a statistic string is appended
to the SDP part of the message. Currently the statistics are
generated with data from the NET and the BTS side.

Put statistics related code into a separete C-File

Generalize the code that it needs only a single connection as
parameter.